refactor: move entity types
