+++ src/hooks/useGames.ts
@@ -1,19 +1,8 @@
 import { useInfiniteQuery } from '@tanstack/react-query'
 import ms from 'ms'
+import { Game } from '../entities/Game'
 import APIClient, { type FetchResponse } from '../services/api-client'
 import useGameQueryStore from '../store/store'
-import { type Platform } from './usePlatforms'
-
-export type Game = {
-  id: number
-  name: string
-  slug: string
-  description_raw: string
-  background_image: string
-  parent_platforms: { platform: Platform }[]
-  metacritic: number
-  rating_top: number
-}
 
 const apiClient = new APIClient<Game>('/games')
