+++ src/hooks/useGenres.ts
@@ -2,12 +2,7 @@ import { useQuery } from '@tanstack/react-query'
 import ms from 'ms'
 import genres from '../data/genres'
 import APIClient from '../services/api-client'
-
-export type Genre = {
-  id: number
-  name: string
-  image_background: string
-}
+import { Genre } from '../entities/Genre'
 
 const apiClient = new APIClient<Genre>('/genres')
