+++ src/hooks/usePlatforms.ts
@@ -1,13 +1,8 @@
 import { useQuery } from '@tanstack/react-query'
 import ms from 'ms'
 import platforms from '../data/platforms'
+import { Platform } from '../entities/Platform'
 import APIClient from '../services/api-client'
-
-export type Platform = {
-  id: number
-  name: string
-  slug: string
-}
 
 const apiClient = new APIClient<Platform>('/platforms')
